Fixed menu class and Added order Class
