+++ src/main/java/com/revature/scoops/util/ConnectionFactory.java
@@ -2,3 +2,4 @@ package com.revature.scoops.util;
 
 public class ConnectionFactory {
 }
+
